requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
Django==4.2.7
django-cors-headers==4.3.1
django-jazzmin==2.6.0
django-ninja==1.0.1
dnspython==2.4.2
email-validator==2.1.0.post1
idna==3.4
Pillow==10.1.0
pydantic==2.5.2
pydantic_core==2.14.5
python-decouple==3.8
sqlparse==0.4.4
typing_extensions==4.8.0
